Use env var in s3 bucket to set initial config

+++ install-dependencies.sh
@@ -1,18 +1,72 @@
 #!/bin/bash
-
+STARTING_DIRECTORY=$(pwd)
+EXPECTED_COMMANDS="jq aws git"
+for expect in $EXPECTED_COMMANDS; do
+  if [[ -z $(which "$expect") ]]; then
+    echo "Please install $expect to continue"
+    exit 1
+  fi
+done
+unset expect
+EXPECTED_PRIMARY_VARIABLES="AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY CONFIG_BUCKET"
+for expect in $EXPECTED_PRIMARY_VARIABLES; do
+  if [[ -z "${!expect}" ]]; then
+    echo "Missing environment variable $expect"
+    exit 1
+  fi
+done
+unset expect
+aws s3 cp $CONFIG_BUCKET .env
+echo "Load environment variables..."
 source .env
+
+EXPECTED_VARIABLES="GIT_TOKEN CONFIG_REPO CONFIG_PATH SLACK_TOKEN WELCOME_CHANNEL"
+for expect in $EXPECTED_VARIABLES; do
+  if [[ -z "${!expect}" ]]; then
+    echo "Missing environment variable $expect"
+    exit 1
+  fi
+done
+unset expect
+mkdir -p tmp
+cd tmp
+CURRENT_PATH=$(pwd)
+git clone $CONFIG_REPO
+if ! [ -f "$CURRENT_PATH/$CONFIG_PATH/admin.json" ]; then
+  echo "Missing admin.json"
+  exit 1
+fi
+if ! [ -f "$CURRENT_PATH/$CONFIG_PATH/config.json" ]; then
+  echo "Missing config.json"
+  exit 1
+fi
+if ! [ -f "$CURRENT_PATH/$CONFIG_PATH/messages.json" ]; then
+  echo "Missing messages.json"
+  exit 1
+fi
+echo "Copying config from repo: $CONFIG_PATH"
+cd $STARTING_DIRECTORY
+#cp "$CURRENT_PATH/$CONFIG_PATH/admin.json" admin.json
+cp "$CURRENT_PATH/$CONFIG_PATH/config.json" config.json
+cp "$CURRENT_PATH/$CONFIG_PATH/messages.json" messages.json
+rm -rf tmp
+
+mkdir -p vendor
+cd vendor
+
 if [[ ! -z $NPM_TOKEN ]]; then
   echo "//registry.npmjs.org/:_authToken=$NPM_TOKEN" > ~/.npmrc
   echo 'registry=https://registry.npmjs.org/' >> ~/.npmrc
 fi
-mkdir -p vendor
-cd ./vendor
 
-DEPENDENCIES=$(cat ../config.json | jq -r '.dependencies[] | " \(.install)\(.source)"' | sed -e 's/\$GITHUB_TOKEN/'"$GITHUB_TOKEN"'/g' | tr '\n' ';')
+echo "Cloning dependencies"
+DEPENDENCIES=$(cat ../config.json | jq -r '.dependencies[] | " \(.install)\(.source)"' | tr '\n' ';')
+# DEPENDENCIES=$(cat ../config.json | jq -r '.dependencies[] | " \(.install)\(.source)"' | sed -e 's/\$GIT_TOKEN/'"$GIT_TOKEN"'/g' | tr '\n' ';')
 echo "Install Dependencies:"
 echo "$(echo $DEPENDENCIES | tr ';' '\n' | sed -e 's/^\s//g')"
 echo "$(eval $(echo $DEPENDENCIES))"
 
+echo "Installing dependencies"
 SETUP=$(cat ../config.json | jq -r '.dependencies[].setup' | tr '\n' ';' | sed -e 's/\;/\; /g')
 echo "Setting up dependencies:"
 echo "$(echo $SETUP | tr ';' '\n' | sed -e 's/^\s//g')"
